+++ test/test_func.py
@@ -1,5 +1,6 @@
 from conftest import self_name
 import pytest
+import inspect
 from webcface.func import Func
 from webcface.func_info import (
     Arg,
@@ -7,6 +8,7 @@ from webcface.func_info import (
     ValType,
     FuncNotFoundError,
     AsyncFuncResult,
+    CallHandle,
 )
 from webcface.field import Field
 from webcface.member import Member
@@ -50,7 +52,7 @@ def test_func_set(data):
 
     f = Func(Field(data, self_name, "c"))
     f.set(
-        lambda x: x + 0.5,
+        lambda x, y: x + 0.5,
         args=[Arg(name="aaaa", type=int), Arg("b")],
         return_type=float,
     )
@@ -109,5 +111,59 @@ def test_func_run(data):
     assert ret.reached
     assert not ret.found
     assert ret.finished
+    assert ret.is_error
+    assert ret.rejection != ""
+
+
+def test_func_handle(data):
+    called = 0
+
+    def f1(a: CallHandle):
+        assert len(a.args) == 4
+        assert a.args[0] == 123
+        assert a.args[1] == 123.45
+        assert a.args[2] == "a"
+        assert a.args[3]
+        nonlocal called
+        called += 1
+        a.respond(123.45)
+
+    assert list(inspect.signature(f1).parameters.values())[0].annotation == CallHandle
+
+    ret = (
+        Func(Field(data, self_name, "a"))
+        .set(
+            f1,
+            args=[
+                Arg(type=int),
+                Arg(type=float),
+                Arg(type=str),
+                Arg(type=bool),
+            ],
+        )
+        .run(123.45, 123.45, "a", True)
+    )
+    assert called == 1
+    assert ret == 123.45
+    ret = Func(Field(data, self_name, "a")).run("123.45", "123.45", "a", 1)
+    assert called == 2
+    assert ret == 123.45
+
+    called = 0
+    ret = Func(Field(data, self_name, "a")).run_async(123.45, 123.45, "a", True)
+    assert ret.reached
+    assert ret.found
+    assert ret.finished
+    assert not ret.is_error
+    assert ret.response == 123.45
+    assert ret.member.name == self_name
+    assert ret.name == "a"
+    assert called == 1
+
+    with pytest.raises(RuntimeError) as e:
+        Func(Field(data, self_name, "a")).run()
+    ret = Func(Field(data, self_name, "a")).run_async()
+    assert ret.reached
+    assert ret.found
     assert ret.is_error
     assert ret.rejection != ""